requirements.txt
Django~=2.2.5
ipython==7.20.0
pymysql~=0.9.3
python-dateutil==2.8.0
mysqlclient==2.0.3
django-cleanup==4.0.0
django_extensions==3.0.9
django-modeladmin-reorder==0.3.1
django_session_timeout==0.1.0
django-storages==1.9.1
django-user-accounts==2.1.0
pillow~=6.2.0
xlsxwriter==1.4.3
mysql-connector==2.2.9